FPGA_Obstacle_Game
DE2 FPGA Board Game: The player interacts with the cube to avoid the oncoming obstacles to beat the level. Any hit from the obstacle resets the level and a death counter will be displayed on the seven segment displays of the board. Initially, there is a main menu screen for the player to press to start the level and a victory screen will show be displayed when the player finally beats the level.